+++ packages/amplify-cli-utils/CHANGELOG.md
@@ -2,8 +2,10 @@
 
  * BREAKING CHANGE: Require Node.js 12.13.0 LTS or newer.
    ([APIGOV-19220](https://jira.axway.com/browse/APIGOV-19220))
+ * refactor: Added `initPlatformAccount()` from `@axway/axway-cli-oum`.
  * feat: Added `telemetry` to help improve Axway products.
    ([APIGOV-19209](https://jira.axway.com/browse/APIGOV-19209))
+ * chore: Updated dependencies.
 
 # v4.4.0 (Jul 30, 2021)